Add nonce-client
