+++ client/Cargo.toml
@@ -46,6 +46,7 @@ solana-faucet = { path = "../faucet", version = "=1.12.0" }
 solana-measure = { path = "../measure", version = "=1.12.0" }
 solana-metrics = { path = "../metrics", version = "=1.12.0" }
 solana-net-utils = { path = "../net-utils", version = "=1.12.0" }
+solana-nonce-client = { path = "../nonce-client", version = "=1.12.0" }
 solana-pubsub-client = { path = "../pubsub-client", version = "=1.12.0" }
 solana-rpc-client = { path = "../rpc-client", version = "=1.12.0" }
 solana-sdk = { path = "../sdk", version = "=1.12.0" }
